Fix #420: support 64-bit integer literals without truncation
Hex/octal literal parsing used unsigned int (32-bit), causing values
above 2^32 to be truncated. Number asString used int (32-bit) for the
integer-form check, so large integers fell into scientific-notation
formatting. Sequence toBase used uintptr_t, which is 32-bit on wasm32.

Switch all three to unsigned long long / long long so integers up to
2^53 (double's exact-integer range) round-trip through literals,
asString, and toBase without loss of precision.

Update NumberTest testAsString: previously expected the buggy
"2.147484e+09" for 2147483648 asString; now expects the correct
decimal "2147483648".

+++ libs/iovm/source/IoSeq_immutable.c
@@ -894,16 +894,16 @@ IO_METHOD(IoSeq, toBase) {
     const char *const table = "0123456789abcdefghijklmnopqrstuvwxyz";
     int base = IoMessage_locals_intArgAt_(m, locals, 0);
     char buf[64], *ptr = buf + 64;
-    uintptr_t n = 0;
+    unsigned long long n = 0;
 
     if (base < 2 || base > 36) {
         IoState_error_(IOSTATE, m, "conversion to base %i not supported", base);
     }
 
 #if defined(_MSC_VER) || defined(__MINGW32__)
-    n = (uintptr_t)strtoull(IoSeq_asCString(self), NULL, 0);
+    n = strtoull(IoSeq_asCString(self), NULL, 0);
 #else
-    n = (uintptr_t)IoSeq_asDouble(self);
+    n = (unsigned long long)IoSeq_asDouble(self);
 #endif
 
     /* Build the converted string backwards. */
